requirements.txt
django>=2.2.4
psycopg2
pillow

django-recaptcha